No need for constraint map refinement
The analysis now properly handles child DSnodes. Maps are created if necessary,
in the event that a map is not created the value itself will be tainted. This is
done so that calls to unknown functions, where only a portion of the structure
is tainted is properly constrained, see sensitivity-test/unknown_fn_struct.